requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.11.0
google-cloud-datastore>=2.18.0
google-adk>=0.1.0
sendgrid>=6.10.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
google-auth>=2.17.0